TP-949 Do not ask confirmation when removing contact info from service
