+++ config/sync/core.entity_form_display.node.service.default.yml
@@ -613,7 +613,7 @@ content:
       display_label: true
     third_party_settings: {  }
   field_contact_info:
-    type: inline_entity_form_complex
+    type: ief_no_confirm_widget
     weight: 9
     region: content
     settings:
@@ -633,7 +633,7 @@ content:
       ief_popup:
         ief_popup_enabled: '1'
   field_contact_info_external:
-    type: inline_entity_form_complex
+    type: ief_no_confirm_widget
     weight: 8
     region: content
     settings:
